Fix style and display
